infinite grid and navigation pt. 1
The grid is now infinite, and you can move around and zoom in/out in the grid. Dragging an object also drags it one dot at a time. This feature is not fully done yet, as later on, the grid's dots will compress and decompress down as you zoom in/out, and the dragging to move around will feel smoother.